% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot}
\alias{plot}
\title{Plot results for sasfclust and sasfclust_cv}
\usage{
plot(mod)
}
\arguments{
\item{mod}{The output of either sasfclust or sasfclust_cv.}
}
\description{
This function plot the estimated cluster mean functions and the classified curves when applied to the output of sasfclust,
whereas, provides a  cross-validation plot when  applied to the output of sasfclust_cv. For the latter, three plots are shown: the first one displays the CV values as a function of  \code{G}, \code{lambda_s} and \code{lambda_l};
the second one displays the CV values as a function of \code{lambda_s} and \code{lambda_l} for \code{G} fixed at its optimal value;
the third one displays the CV values as a function of \code{lambda_l} for \code{G} and \code{lambda_s}   fixed at their optimal value.
}
\examples{
\dontrun{
library(sasfunclust)
train<-simulate_data("Scenario I",n_i=20,var_e = 1,var_b = 0.5^2)
mod<-sasfclust(X=train$X,grid=train$grid,lambda_s = 10^-6,lambda_l =10,G = 2,maxit = 20,q=10)
plot(mod)
}
}
