% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.sasfclust}
\alias{plot.sasfclust}
\title{Plot the results of the Sas-funclust method}
\usage{
\method{plot}{sasfclust}(mod)
}
\arguments{
\item{mod}{The output of   \code{sasfclust}.}
}
\description{
This function povides plots of the estimated cluster mean functions and of the classified curves.
}
\examples{
\dontrun{
library(sasfunclust)
train<-simulate_data("Scenario I",n_i=20,var_e = 1,var_b = 0.5^2)
mod<-sasfclust(X=train$X,grid=train$grid,lambda_s = 10^-6,lambda_l =10,G = 2,maxit = 20,q=10)
plot(mod)
}
}
