% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.sasfclust_cv}
\alias{plot.sasfclust_cv}
\title{Plot the results of the cross-validation for the Sas-funclust method}
\usage{
\method{plot}{sasfclust_cv}(mod)
}
\arguments{
\item{mod}{The output of   \code{sasfclust_cv}.}
}
\description{
This function  provides three plots: the first one displays the CV values as a function of  \code{G}, \code{lambda_s} and \code{lambda_l};
the second one displays the CV values as a function of \code{lambda_s} and \code{lambda_l} for \code{G} fixed at its optimal value;
the third one displays the CV values as a function of \code{lambda_l} for \code{G} and \code{lambda_s}   fixed at their optimal value.
}
\examples{
\dontrun{
library(sasfunclust)
train<-simulate_data("Scenario I",n_i=20,var_e = 1,var_b = 0.5^2)
lambda_s_seq=10^seq(-4,-3)
lambda_l_seq=10^seq(-1,0)
G_seq=2
mod_cv<-sasfclust_cv(X=train$X,grid=train$grid,G_seq=G_seq,
lambda_l_seq = lambda_l_seq,lambda_s_seq =lambda_s_seq,maxit = 20,K_fold = 2,q=10)
plot(mod_cv)
}
}
