% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spfclust_sparse_def.R
\name{sasfclust}
\alias{sasfclust}
\title{FUNCTION_TITLE}
\usage{
sasfclust(
  X = NULL,
  timeindex = NULL,
  curve = NULL,
  grid = NULL,
  q = 30,
  lambda_l = 10,
  lambda_s = 10,
  G = 2,
  tol = 10^-7,
  maxit = 50,
  par_LQA = list(eps_diff = 1e-06, MAX_iter_LQA = 200, eps_LQA = 1e-05),
  plot = F,
  trace = F,
  init = "kmeans",
  varcon = "diagonal",
  lambda_s_ini = NULL
)
}
\arguments{
\item{X}{PARAM_DESCRIPTION, Default: NULL}

\item{timeindex}{PARAM_DESCRIPTION, Default: NULL}

\item{curve}{PARAM_DESCRIPTION, Default: NULL}

\item{grid}{PARAM_DESCRIPTION, Default: NULL}

\item{q}{PARAM_DESCRIPTION, Default: 30}

\item{lambda_l}{PARAM_DESCRIPTION, Default: 10}

\item{lambda_s}{PARAM_DESCRIPTION, Default: 10}

\item{G}{PARAM_DESCRIPTION, Default: 2}

\item{tol}{PARAM_DESCRIPTION, Default: 10^-7}

\item{maxit}{PARAM_DESCRIPTION, Default: 50}

\item{par_LQA}{PARAM_DESCRIPTION, Default: list(eps_diff = 1e-06, MAX_iter_LQA = 200, eps_LQA = 1e-05)}

\item{plot}{PARAM_DESCRIPTION, Default: F}

\item{trace}{PARAM_DESCRIPTION, Default: F}

\item{init}{PARAM_DESCRIPTION, Default: 'kmeans'}

\item{varcon}{PARAM_DESCRIPTION, Default: 'diagonal'}

\item{lambda_s_ini}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Perform the sparse and smooth functional clustering (SaS-Funclust)
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
